handlers/fos_v2/requirements.txt
python-telegram-bot[job-queue]==21.6
gspread==6.1.2
google-auth==2.29.0
apscheduler==3.10.4
